added modal form step for user's very first cycle
