+++ client/src/pages/cycles/CycleForm.js
@@ -3,9 +3,10 @@ import { useForm } from 'react-hook-form';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import Modal from '../../components/modals/Modal';
+import CycleOptions from './CycleOptions';
 import FormStep1 from './FormStep1';
 import FormStep2 from './FormStep2';
-import FormStep3 from './FormStep3';
+import FirstCycle from './FirstCycle';
 
 
 function CycleForm({ currentCycle, toggleModal }) {
@@ -26,8 +27,16 @@ function CycleForm({ currentCycle, toggleModal }) {
   }
   const [formState, setFormState] = useState(initialState)
 
+  // Dyanmic form sequence
+  let forms
+  if (!currentCycle) {
+    // user's first cycle 
+    forms = [FirstCycle, FormStep1, FormStep2]
+  } else {
+    // not the user's first cycle
+    forms = [CycleOptions, FormStep1, FormStep2]
+  }
   // Dynamic component based on step index
-  const forms = [FormStep1, FormStep2, FormStep3]
   const Component = forms[formState.currentStepIndex]
 
   // Repeat Cycle
@@ -44,16 +53,28 @@ function CycleForm({ currentCycle, toggleModal }) {
   const stepIncrement = () => {
     setFormState({
       ...formState,
+      repeat: false,
       currentStepIndex: formState.currentStepIndex + 1,
-  })
+    })
   }
 
   // Decrement step index
   const stepDecrement = () => {
-    setFormState({
+    // this will go back to step 1 from step 3 if user is repeating the cycle weight
+    if (formState.repeat) {
+      setFormState({
         ...formState,
-        currentStepIndex: formState.currentStepIndex - 1,
+        repeat: false,
+        currentStepIndex: formState.currentStepIndex - 2,
     })
+    // this will go back one step from any step if the user is increasing weight
+    } else {
+      setFormState({
+          ...formState,
+          repeat: false,
+          currentStepIndex: formState.currentStepIndex - 1,
+      })
+    }
 }
 
   return (
